feat: move typography to sem
